### 2018-07-24 #806 @Kevin-Mattheus-Moerman
@trallard @moorepants @cybanical @tshu @CELavecchia would you be interested in reviewing this [JOSS](http://joss.theoj.org/about) submission? The review process (once initiated) will take place on GitHub and focuses on [a short paper](https://github.com/openjournals/joss-papers/blob/joss.00806/joss.00806/10.21105.joss.00806.pdf) and the [software](https://github.com/anmuller/CusToM).
