### 2018-07-24 #806 @moorepants
I could review this except that I do not have access to Matlab. Does it run in Octave? If so, I will review.
