### 2018-08-17 #806 @Kevin-Mattheus-Moerman
@trallard are you interested in reviewing this work?
