### 2018-08-30 #806 @Kevin-Mattheus-Moerman
@stavness would you be interested in reviewing this work?
