### 2018-08-30 #806 @Kevin-Mattheus-Moerman
@chrisdembia @sohapouya @modenaxe @demotu would you be interested in reviewing this for the [Journal of Open Source Software](http://joss.theoj.org/)? [The review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#the-joss-review-process) focuses on [the software :floppy_disk:](https://github.com/anmuller/CusToM) as well as a very [short paper :page_facing_up:](https://github.com/openjournals/joss-papers/blob/joss.00806/joss.00806/10.21105.joss.00806.pdf).  Let me know if you are interested or if you have questions.
