### 2018-08-30 #806 @stavness
Hi Kevin,

Thanks for the invitation, but I'm on sabbatical until next July, so I am
trying to limit my admin work.

Ian

On Thu, Aug 30, 2018 at 2:02 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @chrisdembia <https://github.com/chrisdembia> @sohapouya
> <https://github.com/sohapouya> @modenaxe <https://github.com/modenaxe>
> @demotu <https://github.com/demotu> would you be interested in reviewing
> this for the Journal of Open Source Software <http://joss.theoj.org/>? The
> review process
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#the-joss-review-process>
> focuses on the software 💾 <https://github.com/anmuller/CusToM> as well
> as a very short paper 📄
> <https://github.com/openjournals/joss-papers/blob/joss.00806/joss.00806/10.21105.joss.00806.pdf>.
> Let me know if you are interested or if you have questions.
>
