### 2018-09-04 #927 @whedon
[REVIEW]: CusToM: a Matlab toolbox for musculoskeletal simulation
**Submitting author:** @anmuller (<a href="http://orcid.org/0000-0003-2606-1429">Antoine Muller</a>)
**Repository:** <a href="https://github.com/anmuller/CusToM.git" target ="_blank">https://github.com/anmuller/CusToM.git</a>
**Version:** v1.1.0
