### 2018-09-04 #927 @Kevin-Mattheus-Moerman
@modenaxe, @demotu this is where the review will take place. You each have a set of check boxes at the top of this issue to guide you through the review process. Please let me know if you have any questions or check the [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#the-joss-review-process). Thanks! :rocket: :robot:
