### 2018-09-04 #927 @Kevin-Mattheus-Moerman
@anmuller this is where the review process takes place so please keep an eye out for notifications from this issue. Here are some initial comments/questions I have:

1. Is this submission compatible with Octave?
2. Can you please work on adding community guidelines? I recommend adding a CONTRIBUTING.md and a CODE_OF_CONDUCT.md file.
https://help.github.com/articles/setting-guidelines-for-repository-contributors/
You can get a template for the code of conduct file from here: https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
E.g. check out how my personal project has these files:
https://github.com/gibbonCode/GIBBON
