### 2018-09-13 #927 @modenaxe
I just started checking the paper and the material. It's the first time I review for JOSS, so please be patient with me @Kevin-Mattheus-Moerman if I involve you at times :)

**General checks**
- The repository should have a release version number.
- I assume @anmuller has made major contributions to the software, as he is submitting it and the paper cites several publications where he is first author. However, the repository has no history and the functions have no author(s) in the header. @Kevin-Mattheus-Moerman is there some requirement in this respect?

**Functionality**
I installed the toolbox has described in the documentation (just adding the 'Function' folder to the path and ensuring I have all the toolboxes). Then I run 'GenerateParameters.mlapp': the GUI appears but any attempt to use it generates errors. I will follow up in the toolbox repository opening an issue.

**Documentation**
- please mention who your target audience is.
- please list explicitly the dependencies of the toolbox. Currently, in the 'installation instructions' section, it is unclear if the additional Matlab Toolboxes are required for having CusToM working or working 'efficiently', e.g. faster processing.
- the title of section 3 should be modified from 'Application Programming Interface' to 'Graphical User Interface'.
- I think you should specify in the guide that 'GenerateParameters' is in the 'Interface' folder (I assume that is the file you suggest to run in the guide).
I will comment on the rest of the guide after I manage to have the toolbox running.

**Software paper**
Just few comments:
- missing comma after 'body parts'.
- 'thanks to' I would mention explicitly BTK before citing the paper.
- there is a missing reference just after 'inverse dynamics step'.
- 'Antoine Muller et al. 2017': Please make citation style consistent throughout the manuscript.
- Figure 1 is not referenced in the paper and it would be more clear if it was specified explicitly what the external inputs of the boxes are, as in the similar Figure that you have in Muller, Hearing et al. 2017.
