### 2018-09-13 anmuller/CusToM#1 @modenaxe
GenerateParameters.mlapp not working
When, following the documentation, I run (F9) GenerateParameters.mlapp, the GUI appears but any attempt to use it generates errors. I have also tried to define a model using the other tabs and loading parameters from the example, with similar errors in command window.
I am testing your toolbox using MATLAB R2017b (64-bit version).

The complete list of errors when running GenerateParameters.mlapp is the following:

`Undefined function 'Center' for input arguments of type 'char'.

Error in Pelvis (line 69)
Center of mass location with respect to the reference frame

Error in PelvisLowerTrunk (line 26)
[OsteoArticularModel]= Pelvis(OsteoArticularModel,k,Mass,AttachmentPoint);

Error in ModelGeneration (line 42)
[OsteoArticularModel] = ModelParameters.PelvisLowerTrunk(OsteoArticularModel,k,Mass,[]);

Error in PlotModel (line 45)
    [OsteoArticularModel, Markers, Muscles, EnableModel] = ModelGeneration(ModelParameters);

Error in GenerateParameters/actualizePlotModel (line 142)
            evalc('[~,app.NewMarkers,EnableModel] = PlotModel(app.Parameters,app.UIAxes,1,1,1,0,0,0,Anim)');

Error in GenerateParameters/SizeEditFieldValueChanged (line 429)
                app.actualizePlotModel();

Error in GenerateParameters/ActuateApp (line 166)
            app.SizeEditField.Value = OldParameters.Size; app.SizeEditFieldValueChanged();

Error in GenerateParameters/startupFcn (line 307)
            app.ActuateApp(app.Parameters)

Error in GenerateParameters (line 1250)
            runStartupFcn(app, @startupFcn)`
