### 2018-09-13 #927 @modenaxe
I have posted them as an [issue in the repository](https://github.com/anmuller/CusToM/issues/1), let me know if you want more details.
