### 2018-09-13 anmuller/CusToM#1 @anmuller
@modenaxe I think the line 69 of the 'Pelvis' function : "Center of mass location with respect to the reference frame", has to be commented. This sentence is a commentary. Sorry for this inconvenience
