### 2018-09-17 c8a25c5
Correction of the Pelvis function
