### 2018-09-17 180b165
Models correction (refer to issue #1)
