### 2018-09-26 #927 @modenaxe
I am updating my comments after testing the revised toolbox. @anmuller my comments follow exactly the guidelines and checkboxes above.  @Kevin-Mattheus-Moerman this is currently my complete review, excluding the issues that I will post in the repository directly.

**General checks**
- The repository should have a release version number.
- I assume @anmuller has made major contributions to the software, as he is submitting it and the paper cites several publications where he is first author. However, the repository has no history and the functions have no author(s) in the header. I would add the author to the file headers.

**Functionality**
I find challenging to test the functionalities using my own data, mainly due to the documentation, which consists in descriptions of the tabs, but does not help getting started with the toolbox. The included examples do not help much either, as loading the mat file for the analysis and visualisation steps seems to throw errors (which I will report as issues in the repository).

**Documentation**
- please mention who your target audience is.
- please list explicitly the dependencies of the toolbox. Currently, in the 'installation instructions' section, it is unclear if the additional Matlab Toolboxes are required for having CusToM working or working 'efficiently', e.g. faster processing.
- the title of section 3 should be modified from 'Application Programming Interface' to 'Graphical User Interface'.
- In almost all occasions in which you are referring 'saving/running/modifying a file in the corresponding folder' I think you should specify to which file and folder you are referring to, so helping the users. I will post a related issue about reading c3ds in the repository.
- as mentioned above the documentation is descriptive but does not include a real step-by-step example or automated tests to verify the functionality of the code. i think a step-by-step example is crucial here.
- Community guidelines for contributions/reporting issues/seeking support are currently missing.

**Software paper**
Just few comments:
- missing comma after 'body parts'.
- 'set of muscles': I would add a reference to the anatomical datasets you implemented here.
- 'thanks to' I would mention explicitly BTK before citing the paper.
- there is a missing reference just after 'inverse dynamics step'.
- 'Antoine Muller et al. 2017': Please make citation style consistent throughout the manuscript.
- Figure 1 is not referenced in the paper and it would be more clear if it was specified explicitly what the external inputs of the boxes are, as in the similar Figure that you have in Muller, Hearing et al. 2017.
