### 2018-09-26 #927 @nicoguaro
@Kevin-Mattheus-Moerman, I think I can. I have a question, though, is MATLAB necessary for the software or can it be used with Octave? If MATLAB is required, what is the minimum version needed?

Also, from just checking the README in the repo I can't find installation instructions for the package. Or a simple example for "starters". I also suggest to format the README to take advantage of Markdown.
