### 2018-09-27 #927 @Kevin-Mattheus-Moerman
@nicoguaro there is now also a checklist for you at the top of this issue. Yes I believe MATLAB is required. @anmuller can you comment on what version is required?
