### 2018-09-27 #927 @anmuller
It was implemented and tested with Matlab2018a. Also the Symbolic Math, the Optimization and the Parallel Computing toolboxes are necessary
