### 2018-09-28 #927 @anmuller
@nicoguaro Maybe you can use Matlab2017a but I can't ensure you it works.
The installation only consists in adding the folder \CusToM\Functions and its subfolders to the Matlab path. You can check your Matlab version by just running the function "Main" having previously set your current folder as Examples/Cycling or Examples/ROM.
If you are interesting, we created a very quick example of the use of an exemple :
https://drive.google.com/file/d/1_zpWBwQ-S9pBaHbCBJj0qKMDQoVt02PC/view
In the video, all the CusToM folders were added in the path but you have just to add the Functions folders as detailed previously.
