### 2018-09-28 9ffeea7
Bug fixed on video generation
