### 2018-09-28 #927 @nicoguaro
@anmuller, I consider that both the installation instructions and example should be in the repo. Hopefully, in the README.
