### 2018-10-11 e871c55
Addition of instructions in the ReadMe
