+++ README.md
@@ -1,4 +1,4 @@
-## CusToM: a Matlab toolbox for musculoskeletal simulation
+﻿## CusToM: a Matlab toolbox for musculoskeletal simulation
 
 tags:
   - Motion analysis
@@ -29,6 +29,20 @@ According to user choices, the musculoskeletal model generation is achieved than
 For a large set of musculoskeletal models and motion data, CusToM can easily performs all of the analyzes described above. CusToM has been created as a modular tool to let the user being as free and autonomous as possible. The osteoarticular models, set of markers and set of muscles are defined as bricks customizable and adaptable with each other. The design or the modification of a musculoskeletal model is simplified thanks to this modularity. Following the same idea, some methods are defined as adaptable bricks. Testing new cost functions in the optimization schemes, changing performance criteria or creating alternative motion analysis methods can be done in a relatively easy way.
 
 A user interface has been developed to facilitate the data management and the model definition during a given study.
+
+# Installation instructions
+
+CusToM was implemented and tested with the Matlab 2018a version and requires the Symbolic Matlab Toolbox, the Optimization Toolbox and the Parallel Computing Toolbox.
+
+After downloading the main folder named CusToM and placing it in a relevant location on the computer, the installation only consists in adding the folder ...\CusToM\Functions and its subfolders to the Matlab path.
+
+# Data processing examples
+
+Two examples extracted from research works are available in the current release. The first one consists in estimating the lower limbs muscle forces during a cycling motion. The second one consists in predicting the ground reaction forces on a range of motion.
+
+The first example contains an inverse kinematics step, an inverse dynamics step and a muscle forces estimation step. The markers set used is that called M2S in CusToM. External forces applied on each foot were measured with two mobile platforms located on pedals. The subject size is 184 cm and the subject weight is 70 kg. The parameters used for the model generation can be seen by running the GenerateParameters function and by loading the existing ModelParameters (thanks to the Load parameters button). In the same manner, the parameters used for the analysis processing can be seen by running the Analysis function and by loading the existing AnalysisParameters (thanks to the Load parameters button). The analysis can also be directly processed by running the function Main.
+
+The second example contains an inverse kinematics step, a ground reaction forces prediction step and an inverse dynamics step. The subject size is 180 cm and the subject weight us 77.5 kg. The parameters used for model generation can be seen by running the GenerateParameters function and by loading the existing ModelParameters (thanks to the Load parameters button). In the same manner, the parameters used for the analysis processing can be seen by running the Analysis function and by loading the existing AnalysisParameters (thanks to the Load parameters button). The analysis can also be directly processed by running the function Main.
 
 # Acknowledgements
 
