### 2018-10-16 #927 @nicoguaro
I followed the instructions provided in the repo to run ``Main.m`` and get the following result

    Error using load
    Unable to read file 'ModelParameters.mat'. No such file or directory.
