### 2018-10-18 #927 @cpontonn
Dear nicoguaro, did you made the working folder of matlab the one of the example you wanted to run ? If there is no ModelParameters.mat in the folder, therefore you need to create it thanks to the GenerateParameters GUI.

By the way, I am Charles Pontonnier, one of the co-authors of this toolbox. In view of the numerous issues the reviewers raised during the review, we are preparing a lot of additional content to make the handling of the toolbox easier to realize. It may take a bit of time, but we hope that we can commit an amended version of the code and the doc before the end of next week.
