### 2018-10-18 #927 @nicoguaro
@cpontonn, I will wait those changes before any other attempt then.
