+++ Functions/MuscleForces/Optimization/ForcesComputationOpti.m
@@ -18,6 +18,10 @@ function [MuscleForcesComputationResults] = ForcesComputationOpti(filename, Biom
 % Licence
 % Toolbox distributed under 3-Clause BSD License
 %________________________________________________________
+%
+% Authors : Antoine Muller, Charles Pontonnier, Pierre Puchaud and
+% Georges Dumont
+%________________________________________________________
 disp(['Forces Computation (' filename ') ...'])
 
 %% Loading variables
