+++ Functions/Models/Osteoarticular/UpperTrunk/UpperTrunkClavicle.m
@@ -1,0 +1,356 @@
+function [OsteoArticularModel]= UpperTrunkClavicle(OsteoArticularModel,k,Mass,AttachmentPoint)
+% Addition of an upper trunk model
+%   This upper trunk model contains 3 solids (thorax, right clavicle and
+%   left clavicle), exhibits 3 dof for lower trunk / upper trunk joint, 3
+%   dof for each upper trunk / clavicle joint
+%
+%	Based on:
+%	- Damsgaard, M., Rasmussen, J., Christensen, S. T., Surma, E., & De Zee, M., 2006.
+% 	Analysis of musculoskeletal systems in the AnyBody Modeling System. Simulation Modelling Practice and Theory, 14(8), 1100-1111.
+%
+%   INPUT
+%   - OsteoArticularModel: osteo-articular model of an already existing
+%   model (see the Documentation for the structure)
+%   - k: homothety coefficient for the geometrical parameters (defined as
+%   the subject size in cm divided by 180)
+%   - Mass: mass of the solids
+%   - AttachmentPoint: name of the attachment point of the model on the
+%   already existing model (character string)
+%   OUTPUT
+%   - OsteoArticularModel: new osteo-articular model (see the Documentation
+%   for the structure) 
+%________________________________________________________
+%
+% Licence
+% Toolbox distributed under 3-Clause BSD License
+%________________________________________________________
+%
+% Authors : Antoine Muller, Charles Pontonnier, Pierre Puchaud and
+% Georges Dumont
+%________________________________________________________
+list_solid={'UpperTrunk_J1' 'UpperTrunk_J2' 'Thorax' ...
+        'RClavicle_J1' 'RClavicle_J2' 'RClavicle' ...
+        'LClavicle_J1' 'LClavicle_J2' 'LClavicle'};
+ 
+%% solid numbering incremation
+
+s=size(OsteoArticularModel,2)+1;  %#ok<NASGU> % number of the first solid
+for i=1:size(list_solid,2)      % each solid numbering: s_"nom du solide"
+    if i==1
+        eval(strcat('s_',list_solid{i},'=s;'))
+    else
+        eval(strcat('s_',list_solid{i},'=s_',list_solid{i-1},'+1;'))
+    end
+end  
+    
+% find the number of the mother from the attachment point: 'attachment_pt'
+if ~numel(AttachmentPoint)
+    s_mother=0;
+    pos_attachment_pt=[0 0 0]';
+else
+    test=0;
+    for i=1:numel(OsteoArticularModel)
+        for j=1:size(OsteoArticularModel(i).anat_position,1)
+            if strcmp(AttachmentPoint,OsteoArticularModel(i).anat_position{j,1})
+               s_mother=i;
+               pos_attachment_pt=OsteoArticularModel(i).anat_position{j,2}+OsteoArticularModel(s_mother).c;
+               test=1;
+               break
+            end
+        end
+        if i==numel(OsteoArticularModel) && test==0
+            error([AttachmentPoint ' is no existent'])        
+        end       
+    end
+    if OsteoArticularModel(s_mother).child == 0      % if the mother don't have any child
[…]
