+++ Functions/Dynamics/ForwardAllKinematics.m
@@ -13,6 +13,10 @@ function [Human_model] = ForwardAllKinematics(Human_model,j)
 % Licence
 % Toolbox distributed under 3-Clause BSD License
 %________________________________________________________
+%
+% Authors : Antoine Muller, Charles Pontonnier, Pierre Puchaud and
+% Georges Dumont
+%________________________________________________________
 
 if j==0
     return;
