### 2018-11-08 dc9e688
Uptdate Exemples, papers and .md
