+++ Paper/paper.bib
@@ -105,7 +105,7 @@ year = {1999}
 @article{Crowninshield1978,
 title={Use of optimization techniques to predict muscle forces},
 doi = {10.1115/1.3426197},
-author={Crowninshield, Roy D},
+author={Crowninshield, R. D.},
 journal={Journal of Biomechanical Engineering},
 volume={100},
 number={2},
@@ -117,7 +117,7 @@ publisher={American Society of Mechanical Engineers}
 @article{Muller2018,
 title={The MusIC method: a fast and quasi-optimal solution to the muscle forces estimation problem},
 doi = {10.1080/10255842.2018.1429596},  
-author={Muller, Antoine and Pontonnier, Charles and Dumont, Georges},
+author={Muller, A. and Pontonnier, C. and Dumont, G.},
 journal={Computer methods in biomechanics and biomedical engineering},
 volume={21},
 number={2},
@@ -128,7 +128,7 @@ publisher={Taylor \& Francis}
 
 @inproceedings{Muller2017a,
 title={MusIC makes the muscles work together},
-author={Muller, Antoine and Demore, F{\'e}lix and Pontonnier, Charles and Dumont, Georges},
+author={Muller, A. and Demore, F. and Pontonnier, C. and Dumont, G.},
 booktitle={XVI International Symposium on Computer Simulation in Biomechanics},
 pages={2},
 year={2017}
@@ -137,7 +137,7 @@ year={2017}
 @article{Barre2014,
 title={Biomechanical ToolKit: Open-source framework to visualize and process biomechanical data},
 doi = {10.1016/j.cmpb.2014.01.012},  
-author={Barre, Arnaud and Armand, St{\'e}phane},
+author={Barre, A. and Armand, S.},
 journal={Computer methods and programs in biomedicine},
 volume={114},
 number={1},
@@ -145,4 +145,5 @@ pages={80--87},
 year={2014},
 publisher={Elsevier}
 }
+
 
