+++ Examples/SideStep_Processed/Symbolic_function/f3cut.m
@@ -1,0 +1,119 @@
+function [R3cut,p3cut] = f3cut(in1,in2,in3)
+%F3CUT
+%    [R3CUT,P3CUT] = F3CUT(IN1,IN2,IN3)
+
+%    This function was generated by the Symbolic Math Toolbox version 8.1.
+%    25-Oct-2018 19:01:14
+
+R1cut1_1 = in3(1);
+R1cut1_2 = in3(4);
+R1cut1_3 = in3(7);
+R1cut2_1 = in3(2);
+R1cut2_2 = in3(5);
+R1cut2_3 = in3(8);
+R1cut3_1 = in3(3);
+R1cut3_2 = in3(6);
+R1cut3_3 = in3(9);
+p1cut1 = in2(1);
+p1cut2 = in2(2);
+p1cut3 = in2(3);
+q1 = in1(1,:);
+q17 = in1(17,:);
+q18 = in1(18,:);
+q19 = in1(19,:);
+q20 = in1(20,:);
+q21 = in1(21,:);
+t2 = cos(q1);
+t3 = sin(q1);
+t4 = cos(q17);
+t5 = R1cut1_1.*t2;
+t6 = R1cut1_2.*t3;
+t7 = t5+t6;
+t8 = sin(q17);
+t9 = R1cut1_1.*t3;
+t12 = R1cut1_2.*t2;
+t10 = t9-t12;
+t11 = sin(q18);
+t13 = t4.*t10;
+t14 = t7.*t8;
+t15 = t13+t14;
+t16 = cos(q18);
+t17 = cos(q20);
+t18 = R1cut1_3.*t11;
+t30 = t15.*t16;
+t19 = t18-t30;
+t20 = sin(q20);
+t21 = cos(q19);
+t22 = t4.*t7;
+t32 = t8.*t10;
+t23 = t22-t32;
+t24 = sin(q19);
+t25 = t11.*t15;
+t26 = R1cut1_3.*t16;
+t27 = t25+t26;
+t33 = t21.*t23;
+t34 = t24.*t27;
+t28 = t33-t34;
+t29 = cos(q21);
+t31 = t17.*t19;
+t35 = t31-t20.*t28;
+t36 = sin(q21);
+t37 = t19.*t20;
+t38 = t17.*t28;
+t39 = t37+t38;
+t40 = R1cut2_1.*t2;
+t41 = R1cut2_2.*t3;
+t42 = t40+t41;
+t43 = R1cut2_1.*t3;
+t45 = R1cut2_2.*t2;
+t44 = t43-t45;
+t46 = t4.*t44;
+t47 = t8.*t42;
+t48 = t46+t47;
+t49 = R1cut2_3.*t11;
+t58 = t16.*t48;
+t50 = t49-t58;
+t51 = t4.*t42;
+t60 = t8.*t44;
+t52 = t51-t60;
+t53 = t21.*t52;
+t54 = t11.*t48;
+t55 = R1cut2_3.*t16;
+t56 = t54+t55;
+t61 = t24.*t56;
+t57 = t53-t61;
+t59 = t17.*t50;
+t62 = t59-t20.*t57;
+t63 = t20.*t50;
+t64 = t17.*t57;
+t65 = t63+t64;
+t66 = R1cut3_1.*t2;
+t67 = R1cut3_2.*t3;
+t68 = t66+t67;
+t69 = R1cut3_1.*t3;
+t71 = R1cut3_2.*t2;
+t70 = t69-t71;
+t72 = t4.*t70;
+t73 = t8.*t68;
+t74 = t72+t73;
+t75 = R1cut3_3.*t11;
+t84 = t16.*t74;
+t76 = t75-t84;
[…]

[…]
