### 2018-11-08 #927 @anmuller
Thanks for your comments. This is also our first time here and we are sorry if we are not that efficient in answering to everything quickly. Please find the answers of your comments/recommandations. Also a new version of CusToM was committed.

Reviewer 1 : @Kevin-Mattheus-Moerman

Can you please work on adding community guidelines? I recommend adding a CONTRIBUTING.md and a CODE_OF_CONDUCT.md file.
https://help.github.com/articles/setting-guidelines-for-repository-contributors/
You can get a template for the code of conduct file from here : https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
E.g. check out how my personal project has these files:
https://github.com/gibbonCode/GIBBON
Answer: We followed your recommandations and we added, the two .md files: CODE_OF_CONDUCT.md and CONTRIBUTING.md.

Reviewer 2 : @modenaxe

The repository has no history and the functions have no author(s) in the header.
Answer: It has been fixed (authoring in the headers notably).

The documentation is a bit vague about how to load c3d files i.e. apparently they are read if 'in the current folder'. Is that Matlab current working folder? If I copy c3d files on that, and then launch the Analyse GUI, it doesn't seem to read them though, and there is no other obvious way of specifying a folder. I have tried few options and even wondered if it might be due to a lag in the GUI, as in my machine I have noticed it's a bit slow at updating.
Answer: Normally the way to proceed is to make your analysis folder the working folder of matlab. The Analyse GUI is likely to read them from here.

Could you double check if you can run the provided examples without errors? When I start the provided GUIs, load the parameters from the examples folder and then press the 'run' button, Matlab is always complaining and not running the analyses or the visualization.
I can post a more systematic description of the errors if you ensure that the examples are functional in the first place.
Answer: Examples are working to the best of our knowledge: all of the authors checked separately their execution and it worked well for all. We noticed a few cases in which you may have errors: in example 2, if you consider reading external forces from C3D data, it may throw and error since such data is not existing in the C3D. We consider adding an example in which external forces are included in the C3D file to avoid such mistakes. In view of your other comments, we consider writing step-by-step examples with accompanying videos to make it easier to run.

Documentation:
please mention who your target audience is.
Answer: It has been done in the doc. We are  targeting students and researchers in motion analysis, as well as anyone interested in motion analysis (trainers, ergonomists,...)

please list explicitly the dependencies of the toolbox. Currently, in the 'installation instructions' section, it is unclear if the additional Matlab Toolboxes are required for having CusToM working or working 'efficiently', e.g. faster processing.
Answer: The dependencies are mandatory, indeed. We clarified in the doc.

the title of section 3 should be modified from 'Application Programming Interface' to 'Graphical User Interface'.
Answer: Ok, modified
I think you should specify in the guide that 'GenerateParameters' is in the 'Interface' folder (I assume that is the file you suggest to run in the guide). I will comment on the rest of the guide after I manage to have the toolbox running.
Answer: Ok, modified
In almost all occasions in which you are referring 'saving/running/modifying a file in the corresponding folder' I think you should specify to which file and folder you are referring to, so helping the users. I will post a related issue about reading c3ds in the repository.
Answer: Ok, modified.
as mentioned above the documentation is descriptive but does not include a real step-by-step example or automated tests to verify the functionality of the code. i think a step-by-step example is crucial here.
Answer: we agree, and we modified.
Community guidelines for contributions/reporting issues/seeking support are currently missing.
Answer: we handled that the way @Kevin-Mattheus-Moerman proposed.

Software paper:
missing comma after 'body parts'.
'set of muscles': I would add a reference to the anatomical datasets you implemented here.
'thanks to' I would mention explicitly BTK before citing the paper.
there is a missing reference just after 'inverse dynamics step'.
'Antoine Muller et al. 2017': Please make citation style consistent throughout the manuscript.
Figure 1 is not referenced in the paper and it would be more clear if it was specified explicitly what the external inputs of the boxes are, as in the similar Figure that you have in Muller, Hearing et al. 2017.
Answer: We fixed the issues in the paper.

Reviewer 3 : @nicoguaro

From just checking the README in the repo I can't find installation instructions for the package. Or a simple example for "starters". I also suggest to format the README to take advantage of Markdown.
Answer: It has been modified.

I consider that both the installation instructions and example should be in the repo. Hopefully, in the README.
Answer: It has been modified.
