### 2018-11-12 ca7870f
Small modifications to the models and C3D import to support XSENS data
