+++ Functions/Models/Osteoarticular/UpperTrunk/UpperTrunkHead.m
@@ -23,10 +23,6 @@ function [OsteoArticularModel]=UpperTrunkHead(OsteoArticularModel,k,Mass,Attachm
 % Licence
 % Toolbox distributed under 3-Clause BSD License
 %________________________________________________________
-%
-% Authors : Antoine Muller, Charles Pontonnier, Pierre Puchaud and
-% Georges Dumont
-%________________________________________________________
 list_solid={'UpperTrunk_J1' 'UpperTrunk_J2' 'ThoraxHead'};
     
 %% solid numbering incremation
@@ -161,6 +157,8 @@ Thorax_position_set= {...
     'LFHD', LFHD'; ...
     'RBHD', RBHD'; ...
     'LBHD', LBHD'; ...
+    'RMHD', 0.5*(RFHD'+RBHD');
+    'LMHD',0.5*(LFHD'+LBHD');
     'Skull_TopOfHead', Skull_TopOfHead'; ...
     'Skull_NeckNode', Skull_NeckNode'; ...
     'CoM_bary', CoM_bary';...
