+++ Functions/Models/Osteoarticular/Head/Skull.m
@@ -23,10 +23,6 @@ function [OsteoArticularModel]= Skull(OsteoArticularModel,k,Mass,AttachmentPoint
 % Licence
 % Toolbox distributed under 3-Clause BSD License
 %________________________________________________________
-%
-% Authors : Antoine Muller, Charles Pontonnier, Pierre Puchaud and
-% Georges Dumont
-%________________________________________________________
 list_solid={'ThoraxSkull_J1' 'ThoraxSkull_J2' 'Skull'};
        
 %% solid numbering incremation
@@ -81,6 +77,8 @@ Skull_position_set= {...
     'LFHD', k*[0.09 0.095 -0.065]'; ...
     'RBHD', k*[-0.08 0.08 0.07]'; ...
     'LBHD', k*[-0.08 0.08 -0.07]'; ...
+    'RMHD', 0.5*(k*[-0.08 0.08 0.07]'+k*[0.09 0.095 0.065]'); ...
+    'LMHD', 0.5*(k*[-0.08 0.08 -0.07]'+k*[0.09 0.095 -0.065]'); ...
     'Skull_TopOfHead', Skull_TopOfHead'; ...
     'Skull_NeckNode', Skull_NeckNode'; ...
     'GLASS1', [0.13 0.105 0.079]'; ...
