+++ Functions/Models/Osteoarticular/PelvisLowerTrunk/Pelvis.m
@@ -22,10 +22,6 @@ function [OsteoArticularModel]= Pelvis(OsteoArticularModel,k,Mass,AttachmentPoin
 % Licence
 % Toolbox distributed under 3-Clause BSD License
 %________________________________________________________
-%
-% Authors : Antoine Muller, Charles Pontonnier, Pierre Puchaud and
-% Georges Dumont
-%________________________________________________________
 list_solid={'PelvisSacrum'};
     
 %% solid numbering incremation
@@ -96,6 +92,7 @@ Pelvis_position_set= {...
     'Pelvis_HipJointLeftNode', Pelvis_HipJointLeftNode'; ...
     'Pelvis_LowerTrunkNode', Pelvis_SacrumJointNode' + (Sacrum_L5JointNode-Sacrum_PelvisJointNode)'; ...
     'Pelvis_L5JointNode', Pelvis_SacrumJointNode'-Sacrum_L5JointNode'; ...
+    'Pelvis_SacrumJointNode', Pelvis_SacrumJointNode'; ...
     'Pelvis_HipJointsCenterNode', Pelvis_HipJointsCenterNode'; ...
     'CoMPelvis', k*[0 0 0]';...
     };
