+++ Functions/Models/Markers/Marker_set5.m
@@ -16,8 +16,8 @@ s=[s;{  'pRightASI' 'RFWT' {'On';'Off';'On'}; ...
         'pT12SpinalProcess' 'T12' {'Off';'On';'Off'}; ...
         'pT8SpinalProcess' 'T8' {'Off';'On';'Off'}; ...
         'pC7SpinalProcess' 'C7' {'On';'On';'Off'}; ...
-        'pPX' 'STERN' {'On';'Off';'On'}; ...
-        'pIJ' 'ClAV'  {'Off';'Off';'Off'}; ...%not 100% sure
+        'pPX' 'STRN' {'On';'Off';'On'}; ...
+        'pIJ' 'CLAV'  {'Off';'Off';'Off'}; ...%not 100% sure
         'pTopOfHead' 'Skull_TopOfHead' {'Off';'Off';'On'}; ...
         'pRightAuricularis' 'RMHD' {'Off';'Off';'On'}; ...
         'pLeftAuricularis' 'LMHD' {'Off';'Off';'On'};...
