### 2018-11-13 23550dc
Authoring for XSENS markerset
