### 2018-11-13 056bc43
Added authors names again on models functions
