### 2018-12-06 fcf5eab
test presence of BiomechanicalModel.mat only in current folder
