### 2018-12-13 #927 @modenaxe
@anmuller thank you for replying to my comments. I have tested the updated toolbox.
Few points:
1) you still don't have a version for your current toolbox. I would just tag the current version using git.
2) the toolbox seems to work also with MATLABR2017b, is there a reason why the 'Installation.m' file requires R2018a?
3) the toolbox works, but when I run the analyses from the GUI, MATLAB does not appear as 'Busy' and it takes several minutes for the computations. Maybe this is due to the Matlab version? (the use of computational resources, i.e. CPU and memory is low).
For the rest all seems to work, well done!
If you address this last comments for me the paper is accepted.
