### 2018-12-14 #927 @anmuller
@modenaxe Thank you for your review and for the different comments.

Please find the answers of the points you mentionned:
1. We tagged the current version of the toolbox.
2. Indeed, the toolbox seems to work with Matlab2017b but as the development was done with the 2018a version, we can not guarantee that the code could be run on previous versions. Also, the graphical use interfaces could only be modified on Matlab2018a.
3. It is normal that the computation takes several minutes (the generation of the biomechanical model and the different motion analysis steps may take some time). However, we do not know why Matlab does not appear 'Busy' during these computations.

We hope this has adressed your points.
