### 2018-12-14 #927 @demotu
I tested CusToM using matlab 2018b in a windows machine. With the help of the documentation and the tutorials on youtube, i was able to run the three examples that come with the software.
This software is for a very specialized public but it works as described.
I think after all the improvements the authors have made this paper is ready to be accepted.
