### 2018-12-26 #927 @nicoguaro
I tried to run the example described in section 4.1 of the documentation in the following Linux system:

    CentOS Linux release 7.4.1708

and obtained the following error:

    MathWorks::System::IUserException [Error using matlab.internal.webwindow (line 312)
    The MATLABWindow application failed to launch because of missing libraries. Install the packages which contain the following libraries:

	    libXss.so.1

### Other details
Other minor details are:

 - The **Statement of need** is still missing in the repository.
 - The summary can use Markdown formatting for the cited references.
 - There isn't a version number in the repository.
