### 2019-01-04 #927 @Kevin-Mattheus-Moerman
@nicoguaro thanks for these comments. @anmuller can you work on these points?
