### 2019-01-04 #927 @Kevin-Mattheus-Moerman
The version number can be set/fixed properly at a later stage (e.g. we assign the correct one if the paper is accepted).
