### 2019-01-09 e8881a7
Update README and Documentation.
