+++ README.md
@@ -7,6 +7,7 @@
 
 # Table of contents
 - [Authors](#Authors)
+- [Statement of Need](#StatementofNeed)
 - [Summary](#Summary)  
 - [Installation instructions](#Installationinstructions)
 - [Data processing examples](#Dataprocessingexamples)
@@ -14,7 +15,8 @@
 - [Contributing](#Contributing)  
 - [Code of conduct](#CodeOfConduct)  
 - [License](#License)
-- [Acknowledgements](#Acknowledgements)  
+- [Acknowledgements](#Acknowledgements)
+- [Bibliography](#Bibliography)
 
 Tags:
   - Motion analysis
@@ -37,32 +39,49 @@ Tags:
    
 date: 2 July 2018
 
+# Statement of Need <a name="StatementofNeed"></a>
+
+Inverse dynamics based musculoskeletal analysis aims at calculating biomechanical quantities to understand motion from joint kinematics to muscle forces.
+Generic models generally based on cadaveric templates are used as an input. 
+These generic models are based on three layers modelisation. 
+The first one is geometrical defining the polyarticulated rigid body system and the kinematic joints. 
+The inertial layer defines mass, centers of mass position and inertia matrix of rigid body of the polyarticulated system. 
+And, the muscle layer defines the muscle paths and force generation behaviours of muscles. 
+This generic models are then calibrated through multiple calibration steps to be subject-specific base on motion capture data. 
+Finally the subject-specific model is used to understand recorded motions of the subject.
+
+However, musculoskeletal simulation requires high computational cost. Editing and assembling features to modify generic models. 
+Subject-specific calibrations and multiple simulations are required to compute subject-specific quantities on recorded trials. 
+And current available musculoskeletal softwares are heavy and requires expertise like SIMM (MusculoGraphics, Inc., Santa Rosa, CA), Anybody (Anybody Technology, Aalborg, Denmark) and OpenSim OpenSim (Simtk, Stanford, CA). 
+Moreover, SIMM and Anybody are commercial softwares and, OpenSim is a freely available software but main algorithms source codes are not available. 
+That's why, there was a need in developping a complete open-source software for musculoskeletal simulation. 
+The source code was developped with Matlab to allow many researchers to understand and contribute to the code.
+
 # Summary <a name="Summary"></a>
 
 Customizable Toolbox for Musculoskeletal simulation (CusToM) is a MATLAB toolbox aimed 
-at performing inverse dynamics based musculoskeletal analyzes [@Erdemir2007]. This type 
+at performing inverse dynamics based musculoskeletal analyzes (Erdemir et al., 2007). This type 
 of analysis is essential to access mechanical quantities of human motion in different 
 fields such as clinics, ergonomics and sports. CusToM exhibits several features. It 
 can generate a personalized musculoskeletal model, and can solve from motion capture 
 data inverse kinematics, external forces estimation, inverse dynamics and muscle forces
- estimation problems as in various musculoskeletal simulation software [@Damsgaard2006]
- [@Delp2007].
+ estimation problems as in various musculoskeletal simulation software (Damsgaard et al. 2006, Delp et al. 2007).
 
 According to user choices, the musculoskeletal model generation is achieved thanks to 
-libraries containing pre-registered models [@Muller2015b]. These models consist of body
+libraries containing pre-registered models (Muller et al. 2015). These. These models consist of body
 parts osteoarticular models, set of markers or set of muscles to be combined together.
[…]
