### 2019-01-09 #927 @Ipuch
Thank you @nicoguaro.

Unfortunately, the library have been developped only on Windows 10. We added the information in the documentation and in the README.md.

> obtained the following error:
>`MathWorks::System::IUserExceptionError using matlab.internal.webwindow (line 312)
The MATLABWindow application failed to launch because of missing libraries. Install the packages which contain the following libraries: libXss.so.1`

I found two hypotheses to your issue :

- https://fr.mathworks.com/matlabcentral/answers/385952-why-do-i-see-error-while-loading-shared-libraries-libxss-so-1-when-starting-matlab-on-a-linux-mac
- CusToM was not developped on MacosX and Linux. For MacosX, it could be necessary to download the source files of BTK and to compile and install BTK accordingly with your device. You would need to download [btk-core-0.3.0\_src.zip](https://code.google.com/archive/p/b-tk/downloads). (This information was added to the documentation and README.md)

Here is the statement of need we added to the README.md before the summary and to the documentation:

Inverse dynamics based musculoskeletal analysis aims at calculating biomechanical quantities to understand motion from joint kinematics to muscle forces. Generic models generally based on cadaveric templates are used as an input. These generic models are based on three layers modelisation.
The first one is geometrical defining the polyarticulated rigid body system and the kinematic joints. The inertial layer defines mass, centers of mass position and inertia matrix of rigid body of the polyarticulated system. And, the muscle layer defines the muscle paths and force generation behaviours of muscles.
This generic models are then calibrated through multiple calibration steps to be subject-specific base on motion capture data. Finally the subject-specific model is used to understand recorded motions of the subject.

However, musculoskeletal simulation requires high computational cost. Editing and assembling features to modify generic models. Subject-specific calibrations and multiple simulations are required to compute subject-specific quantities on recorded trials. And current available musculoskeletal softwares are heavy and requires expertise like SIMM (MusculoGraphics, Inc., Santa Rosa, CA), Anybody (Anybody Technology, Aalborg, Denmark) and OpenSim OpenSim (Simtk, Stanford, CA). Moreover, SIMM and Anybody are commercial softwares and, OpenSim is a freely available software but main algorithms source codes are not available. That's why, there was a need in developping a complete open-source software for musculoskeletal simulation. The source code was developped with Matlab to allow many researchers to understand and contribute to the code.

References of the summary in the README.md were added to a bibliography at the end of the document. The formatting of the citations were revised.

The version number will be set later stage as recommended by @Kevin-Mattheus-Moerman
