### 2019-01-09 b3bbe58
minor edit on names of the bibliography in the README
