### 2019-01-12 #927 @nicoguaro
@Kevin-Mattheus-Moerman, currently I don't have access to MATLAB 2018 under Windows. So I can't attest for the functionality and performance. Besides that I am content with how they addressed my comments.
