### 2019-01-17 #927 @cpontonn
Thank you for your comments. @Kevin-Mattheus-Moerman , since we do not consider developing a Linux compatible version for now, what is the next step to validate the review ? Thanks in advance.
