### 2019-01-17 #927 @Kevin-Mattheus-Moerman
@cpontonn
I have two minor editorial comments about the paper:
- [ ] Please change this sentence to include the highlighted word: "The motion capture
data (contained in a c3d file) are extracted with `the` Biomechanical ToolKit (BTK) (Barre &
Armand, 2014)"
- [ ] Perhaps fix the use of "Last" in this sentence: "Last, muscle forces are estimated.". Consider using "Lastly" or "Finally" instead.

Please implement these changes and use `@whedon generate pdf` here to update the paper.
Thanks.
