### 2019-01-17 #927 @nicoguaro
@Kevin-Mattheus-Moerman, the paper looks OK to my eyes!
