### 2019-01-17 #927 @Kevin-Mattheus-Moerman
@cpontonn I will now recommend that this work is accepted in JOSS.

@modenaxe, @demotu, @nicoguaro thanks for reviewing this work!! :rocket: :robot:
