### 2019-01-17 #927 @Kevin-Mattheus-Moerman
@cpontonn at this point please provide us with:

- [ ] The DOI of the archived version of the final reviewed software. You can create an archived version with a service like [ZENODO](https://zenodo.org/) (see also the useful steps described [here](https://guides.github.com/activities/citable-code/)).

- [ ] Please provide the version/release number for the archived version of the software that was reviewed.
