### 2019-01-18 #927 @anmuller
@Kevin-Mattheus-Moerman, this is the Zenedo DOI : 10.5281/zenodo.2543645.
Also the version for the archived version of the software is v1.1.0.
Thanks
