### 2019-01-18 #927 @labarba
I see that @nicoguaro has two un-checked items, but also see his comment:

> currently I don't have access to MATLAB 2018 under Windows. So I can't attest for the functionality and performance. Besides that I am content with how they addressed my comments.

Given that the submission has three reviewers, I will assume that the @Kevin-Mattheus-Moerman's recommendation to publish stands.
