### 2019-01-18 #927 @labarba
@anmuller — Please edit the metadata of the Zenodo deposit (title and author list) to match the JOSS paper. Thanks!
