### 2019-01-18 #927 @labarba
Finally, journal titles are capitalized in English-language publication. You're mixing English and French conventions. Please capitalize throughout.
