### 2019-01-18 #927 @Kevin-Mattheus-Moerman
@labarba correct there was an issue with Linux and this limitation was added to the descriptions/documentation. Rather than removing the reviewer I thought it best to leave these statements for the record.
Thanks for addressing those additional editorial points that I missed.
