### 2019-01-18 6ce57d6
Minor changes on paper.bib
