+++ Paper/paper.bib
@@ -34,7 +34,7 @@ year = {2017}
 
 @book{Featherstone2008,
 author = {Featherstone, R.},
-doi = {10.1007/978-0-387-74315-8},
+doi = {10.1007/978-1-4899-7560-7},
 pages = {272},
 title = {{Rigid Body Dynamics Algorithms}},
 year = {2008}
@@ -61,7 +61,7 @@ year = {2016}
 @article{Delp2007,
 author = {Delp, S. L. and Anderson, F. C and Arnold, A. S and Loan, P. and Habib, A. and John, C. T. and Guendelman, E. and Thelen, D. G.},
 doi = {10.1109/TBME.2007.901024},
-journal = {IEEE transactions on bio-medical engineering},
+journal = {IEEE Transactions on Biomedical Engineering},
 number = {11},
 pages = {1940--1950},
 title = {{OpenSim: Open source to create and analyze dynamic simulations of movement}},
@@ -71,7 +71,7 @@ year = {2007}
 
 @inproceedings{Muller2015b,
 author = {Muller, A. and Pontonnier, C. and Germain, C. and Dumont, G.},
-booktitle = {Proceedings of the 40{\`{e}}me Congr{\`{e}}s de la Soci{\'{e}}t{\'{e}} de Biom{\'{e}}canique, Computer methods in biomechanics and biomedical engineering},
+booktitle = {Proceedings of the 40{\`{e}}me Congr{\`{e}}s de la Soci{\'{e}}t{\'{e}} de Biom{\'{e}}canique, Computer Methods in Biomechanics and Biomedical Engineering},
 doi = {10.1080/10255842.2015.1069599},
 number = {sup1},
 pages = {2008--2009},
@@ -94,7 +94,7 @@ year = {2007}
 @article{Lu1999,
 author = {Lu, T. W. and O'connor, J. J.},
 doi = {10.1016/S0021-9290(98)00158-4},
-journal = {Journal of biomechanics},
+journal = {Journal of Biomechanics},
 number = {2},
 pages = {129--134},
 title = {{Bone position estimation from skin marker co-ordinates using global optimisation with joint constraints}},
@@ -118,7 +118,7 @@ publisher={American Society of Mechanical Engineers}
 title={The MusIC method: a fast and quasi-optimal solution to the muscle forces estimation problem},
 doi = {10.1080/10255842.2018.1429596},  
 author={Muller, A. and Pontonnier, C. and Dumont, G.},
-journal={Computer methods in biomechanics and biomedical engineering},
+journal={Computer Methods in Biomechanics and Biomedical Engineering},
 volume={21},
 number={2},
 pages={149--160},
@@ -138,7 +138,7 @@ year={2017}
 title={Biomechanical ToolKit: Open-source framework to visualize and process biomechanical data},
 doi = {10.1016/j.cmpb.2014.01.012},  
 author={Barre, A. and Armand, S.},
-journal={Computer methods and programs in biomedicine},
+journal={Computer Methods and Programs in Biomedicine},
 volume={114},
 number={1},
 pages={80--87},
