+++ Paper/paper.bib
@@ -36,7 +36,7 @@ year = {2017}
 author = {Featherstone, R.},
 doi = {10.1007/978-1-4899-7560-7},
 pages = {272},
-title = {{Rigid Body Dynamics Algorithms}},
+title = {{Rigid body dynamics algorithms}},
 year = {2008}
 }
 
