### 2019-01-21 d1e7ad8
Minor changes in paper.bib
