+++ Paper/paper.bib
@@ -16,7 +16,7 @@ year = {2014}
 
 @inproceedings{Muller2017c,
 author = {Muller, A. and Haering, D. and Pontonnier, C. and Dumont, G.},
-booktitle = {Proceedings of the 23{\`{e}}me Congr{\`{e}}s Fran{\c{c}}ais de M{\'{e}}canique},
+booktitle = {Proceedings of the 23{\`{e}}me {C}ongr{\`{e}}s {F}ran{\c{c}}ais de {M}{\'{e}}canique},
 title = {{Non-invasive techniques for musculoskeletal model calibration}},
 year = {2017}
 }
@@ -53,7 +53,7 @@ year = {2006}
 
 @inproceedings{Muller2015a,
 author = {Muller, A. and Germain, C. and Pontonnier, C. and Dumont, G.},
-booktitle = {Proceedings of the 33rd International Society of Biomechanics in Sports},
+booktitle = {Proceedings of the 33rd {I}nternational {S}ociety of {B}iomechanics in {S}ports},
 title = {{A simple method to calibrate kinematical invariants: application to overhead throwing}},
 year = {2016}
 }
@@ -71,7 +71,7 @@ year = {2007}
 
 @inproceedings{Muller2015b,
 author = {Muller, A. and Pontonnier, C. and Germain, C. and Dumont, G.},
-booktitle = {Proceedings of the 40{\`{e}}me Congr{\`{e}}s de la Soci{\'{e}}t{\'{e}} de Biom{\'{e}}canique, Computer Methods in Biomechanics and Biomedical Engineering},
+booktitle = {Proceedings of the 40{\`{e}}me {C}ongr{\`{e}}s de la {S}oci{\'{e}}t{\'{e}} de {B}iom{\'{e}}canique, {C}omputer {M}ethods in {B}iomechanics and {B}iomedical {E}ngineering},
 doi = {10.1080/10255842.2015.1069599},
 number = {sup1},
 pages = {2008--2009},
@@ -129,7 +129,7 @@ publisher={Taylor \& Francis}
 @inproceedings{Muller2017a,
 title={MusIC makes the muscles work together},
 author={Muller, A. and Demore, F. and Pontonnier, C. and Dumont, G.},
-booktitle={XVI International Symposium on Computer Simulation in Biomechanics},
+booktitle={XVI {I}nternational {S}ymposium on {C}omputer {S}imulation in {B}iomechanics},
 pages={2},
 year={2017}
 }
