### 2019-01-21 #927 @anmuller
@labarba thanks for your comments.
Actually the DOI for the reference Featherstone was wrong and it was corrected. The other changes you mentionned were done. We hope that everything is good.
