### 2019-01-21 7b4a449
Minor changes on the paper
