+++ Paper/paper.md
@@ -33,18 +33,19 @@ bibliography: paper.bib
 
 # Summary
 
-Customizable Toolbox for Musculoskeletal simulation (CusToM) is a MATLAB toolbox aimed at performing inverse dynamics based musculoskeletal analyzes [@Erdemir2007]. This type of analysis is essential to access mechanical quantities of human motion in different fields such as clinics, ergonomics and sports. CusToM exhibits several features. It can generate a personalized musculoskeletal model, and can solve from motion capture data inverse kinematics, external forces estimation, inverse dynamics and muscle forces estimation problems as in various musculoskeletal simulation software [@Damsgaard2006] [@Delp2007] (Figure \ref{Pipeline}).
+Customizable Toolbox for Musculoskeletal simulation (CusToM) is a MATLAB toolbox aimed at performing inverse-dynamics-based musculoskeletal analyses [@Erdemir2007]. This type of analysis is essential to access mechanical quantities of human motion in different fields such as clinics, ergonomics and sports. CusToM exhibits several features. It can generate a personalized musculoskeletal model, and can solve from motion-capture data inverse kinematics, external forces estimation, inverse dynamics and muscle forces estimation problems as in various musculoskeletal simulation software [@Damsgaard2006; @Delp2007] (Figure \ref{Pipeline}).
 
-According to user choices, the musculoskeletal model generation is achieved by using libraries containing pre-registered models [@Muller2015b]. These models consist of body parts of osteoarticular models, set of markers or set of muscles to be combined together. From an anthropometric based model, the geometric, inertial and muscular parameters are calibrated to fit the size and mass of the subject to be analyzed [@Muller2015a] [@Muller2017] [@Muller2017c]. The motion capture data (contained in a c3d file) are extracted with the Biomechanical ToolKit (BTK) [@Barre2014]. Then, the inverse kinematics step computes joint coordinates trajectories against time [@Lu1999]. Then, joint torques are computed in an inverse dynamics step [@Featherstone2008]. To this end, external forces applied to the subject have to be known. They may be directly extracted from experimental data (as platform forces) or be estimated from motion data by using the equations of motion in an optimization scheme [@Fluit2014]. Lastly, muscle forces are estimated. It consists in finding a repartition of muscle forces respecting the joint torques and representing the central nervous system strategy [@Crowninshield1978] [@Muller2017a] [@Muller2018].
+According to user choices, the musculoskeletal model generation is achieved by using libraries containing pre-registered models [@Muller2015b]. These models consist of body parts of osteoarticular models, set of markers or set of muscles to be combined together. From an anthropometric based model, the geometric, inertial and muscular parameters are calibrated to fit the size and mass of the subject to be analyzed [@Muller2015a; @Muller2017; @Muller2017c]. The motion capture data (contained in a c3d file) are extracted with the Biomechanical ToolKit (BTK) [@Barre2014]. The inverse kinematics step computes joint coordinates trajectories against time [@Lu1999]. Joint torques are computed in an inverse dynamics step [@Featherstone2008]. To this end, external forces applied to the subject have to be known. They may be directly extracted from experimental data (as platform forces) or be estimated from motion data by using the equations of motion in an optimization scheme [@Fluit2014]. Lastly, muscle forces are estimated. It consists of finding a repartition of muscle forces respecting the joint torques and representing the central nervous system strategy [@Crowninshield1978; @Muller2017a; @Muller2018].
 
 ![CusToM pipeline\label{Pipeline}](Pipeline.pdf)
 
[…]
