### 2019-01-21 #927 @anmuller
@labarba thanks again for the comments and for the advices. The changes were done.
