### 2019-01-21 #927 @labarba
Congratulations, @anmuller, your paper is published in JOSS! 🎉

Thank you, gracias, merci to your handling editor, @Kevin-Mattheus-Moerman, and your reviewers, @modenaxe, @demotu, @nicoguaro 🙏
